Updated mesh bvh construction splitting direction.
Was using largest phys dim as the dim to split a ref box. For volume meshes this is somewhere between sensible and random, but for surface meshes it is not even defined.
Changed to use the largest ref dim.
We still want a geometry based heuristic, so that will have to be a TODO.