New filter to project the mesh to a higher degree and encapsulate in a new dataset.
In preparation to create a degree-raising furnace.

Note that you would not really want to use the resulting mesh
for anything other than improved bounds or improving the bvh.
The topology is currently destroyed (shared vertices are no
longer shared). Also the evaluation will contain a small
amount of error, better to stick to original.