Added external eval_d for General Hex
The primary difference is the interface change, as we migrate out of the Element Object abstraction.

Implementation-wise, this version extends the Horner's rule as used in eval_d_edge/eval_d_face,
instead of as used in pos_tensor_element.tcc. I.e., no left/right to keep track of.